add main service logic

+++ src/docker_postgres_init.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS persons (
   id SERIAL,
   name varchar(250) NOT NULL,
-  features integer[256] NOT NULL,
+  features double precision[256] NOT NULL,
   PRIMARY KEY (id)
 );
